+++ plugins/native/module/vtkF3DQuakeMDLImporterConstants.h
@@ -1,0 +1,7 @@
+#ifndef vtkF3DQuakeMDLImporterConstants_h
+#define vtkF3DQuakeMDLImporterConstants_h
+
+extern const unsigned char F3DMDLDefaultColorMap[256][3];
+extern const float F3DMDLNormalVectors[162][3];
+
+#endif
