--- plugins/native/module/vtkQuakeMDLImporter.h.rej
@@ -1,17 +1,0 @@
-diff a/plugins/native/module/vtkQuakeMDLImporter.h b/plugins/native/module/vtkQuakeMDLImporter.h	(rejected hunks)
-@@ -5,14 +5,13 @@
- 
- #ifndef vtkQuakeMDLImporter_h
- #define vtkQuakeMDLImporter_h
--#include<vtkImporter.h>
-+#include <vtkImporter.h>
- 
- class vtkQuakeMDLImporter : public vtkImporter
- {
- public:
-   static vtkQuakeMDLImporter* New();
-   vtkTypeMacro(vtkQuakeMDLImporter, vtkImporter);
--
- 
-   void PrintSelf(ostream& os, vtkIndent indent) override;
-
