+++ testing/baselines/TestQuakeMDL.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7b8ebba71f1475f7bf2382b3736ff82147519b71a9d290f3ac43bee4adcb60ed
+size 47901
